requirements.txt
flask==2.0.1
python-chess==1.9.0
gunicorn==20.1.0
Werkzeug==2.0.1 